LayoutTests/imported/w3c/web-platform-tests/css/css-color-adjust/rendering/dark-color-scheme/color-scheme-iframe-preferred-change-expected.txt



FAIL Preferred color-scheme of iframe document changed through iframe element color-scheme assert_equals: expected "rgb(128, 0, 128)" but got "rgb(0, 0, 255)"
